Refactor: onSuccess, onError 시 toast 출력 코드 ImageMenuBar 파일로 이동
